+++ dev.config.js
@@ -3,15 +3,32 @@ const path = require('path');
 module.exports = {
   mode: 'development',
   entry: {
-    test: './src/test.js',
-    mint: './src/mint.js',
-    'stats-tester': './src/stats-tester.js',
-    'stats-fct': './src/stats-fct.js',
+    test: './src/test.ts',
+    mint: './src/mint.ts',
+    'stats-tester': './src/stats-tester.ts',
+    'stats-fct': './src/stats-fct.ts',
   },
   output: {
     filename: '[name].js',
     path: __dirname + '/dist',
     publicPath: '/dist/',
+  },
+  // Enable sourcemaps for debugging webpack's output.
+  devtool: "source-map",
+  resolve: {
+    extensions: [".ts", ".js"],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts$/,
+        loader: 'ts-loader',
+      },
+      {
+        test: /\.js$/,
+        loader: "source-map-loader",
+      },
+    ],
   },
   devServer: {
     static: {
