+++ package.json
@@ -4,9 +4,12 @@
   "description": "Project description",
   "main": "index.js",
   "private": "true",
-  "dependencies": {
-  },
+  "dependencies": {},
   "devDependencies": {
+    "@types/p5": "^1.3.1",
+    "source-map-loader": "^3.0.0",
+    "ts-loader": "^9.2.6",
+    "typescript": "^4.5.2",
     "webpack": "^5.37.0",
     "webpack-cli": "^4.7.0",
     "webpack-dev-server": "^4.5.0"
