--- src/stats-tester.js
@@ -1,3 +1,0 @@
-import StatsFct from './stats-fct';
-
-console.log(features);
