+++ src/stats-tester.ts
@@ -1,0 +1,4 @@
+import './stats-fct';
+
+declare const features: any;
+console.log(features);
