+++ src/types/global.d.ts
@@ -1,0 +1,2 @@
+declare var features: any;
+declare var tokenData: { tokenId: string; hash: string; };
