+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "outDir": "./dist/",
+    "sourceMap": true,
+    "allowJs": true,
+    "module": "es6",
+    "target": "es6",
+    "moduleResolution": "node",
+    "types": ["p5"]
+  },
+  "include": ["./src/**/*", "./src/types/global.d.ts", "./global.d.ts"]
+}
